refactor database backup to use existing slide iterator and workflows class
